refactor(sdk,sdkx)!: move graph agent factory to sdk/graph/config
